Refactor particle swarm optimization module
- Introduced a new `Swarm` module to encapsulate swarm-related functionality, including particle management and boundary handling.
- Updated the `PSO` struct to utilize the new `Swarm` module, modifying methods to accept bounds for particle initialization and updates.
- Refactored `SwarmStatus` to streamline particle retrieval and best position evaluation, removing redundant methods.
- Adjusted the `Solver` trait to accommodate bounds in initialization, updates, and summarization methods.
