+++ src/core/minimizer.rs
@@ -4,19 +4,25 @@ use parking_lot::RwLock;
 
 use crate::traits::{AbortSignal, CostFunction, Observer, Solver, Status};
 
-use super::{Config, NopAbortSignal, Summary};
+use super::{Bound, NopAbortSignal, Summary};
 
+const DEFAULT_MAX_STEPS: usize = 4000;
 /// The main struct used for running [`Solver`]s on [`CostFunction`]s.
 pub struct Minimizer<S, U, E> {
     /// The [`Status`] of the [`Solver`], usually read after minimization.
     pub status: S,
+    /// The [`Summary`] of the [`Solver`], usually read after minimization.
+    pub result: Option<Summary>,
+
     solver: Box<dyn Solver<S, U, E>>,
     observers: Vec<Arc<RwLock<dyn Observer<S, U>>>>,
     abort_signal: Box<dyn AbortSignal>,
-    config: Config,
     user_data: U,
-    /// The [`Summary`] of the [`Solver`], usually read after minimization.
-    pub result: Option<Summary>,
+
+    dimension: usize,
+    bounds: Option<Vec<Bound>>,
+    parameter_names: Option<Vec<String>>,
+    max_steps: usize,
 }
 
 impl<S: Status, U: Default, E> Minimizer<S, U, E> {
@@ -25,10 +31,10 @@ impl<S: Status, U: Default, E> Minimizer<S, U, E> {
     pub fn new(solver: Box<dyn Solver<S, U, E>>, dimension: usize) -> Self {
         Self {
             status: S::default(),
-            config: Config {
-                dimension,
-                ..Default::default()
-            },
+            dimension,
+            bounds: None,
+            parameter_names: None,
+            max_steps: DEFAULT_MAX_STEPS,
             solver,
             observers: Vec::default(),
             abort_signal: NopAbortSignal.boxed(),
@@ -56,15 +62,6 @@ impl<S: Status, U: Default, E> Minimizer<S, U, E> {
         self
     }
 
-    /// Edit the [`Config`] of the [`Minimizer`].
-    pub fn on_config<F>(&mut self, mut f: F) -> &mut Self
-    where
-        F: FnMut(&mut Config) -> &mut Config,
-    {
-        f(&mut self.config);
-        self
-    }
-
     /// Edit the [`Status`] of the [`Minimizer`].
     pub fn on_status<F>(&mut self, mut f: F) -> &mut Self
     where
@@ -74,6 +71,42 @@ impl<S: Status, U: Default, E> Minimizer<S, U, E> {
         self
     }
 
+    /// Sets all [`Bound`]s of the [`Config`] used by the [`Solver`](crate::traits::Solver). This can be [`None`] for an unbounded problem, or
+    /// [`Some`] [`Vec<(T, T)>`] with length equal to the number of free parameters. Individual
+    /// upper or lower bounds can be unbounded by setting them equal to `T::infinity()` or
+    /// `T::neg_infinity()` (e.g. `f64::INFINITY` and `f64::NEG_INFINITY`).
+    ///
+    /// # Panics
+    ///
+    /// This function will panic if the number of bounds is not equal to the number of free
+    /// parameters.
+    pub fn with_bounds<I: IntoIterator<Item = B>, B: Into<Bound>>(
+        &mut self,
+        bounds: I,
+    ) -> &mut Self {
+        let bounds = bounds.into_iter().map(Into::into).collect::<Vec<Bound>>();
+        assert!(bounds.len() == self.dimension);
+        self.bounds = Some(bounds);
+        self
+    }
+
+    /// Sets the names of the parameters. This is only used for printing and debugging purposes.
+    ///
+    /// # Panics
+    ///
+    /// This function will panic if the number of bounds is not equal to the number of free
+    /// parameters.
+    pub fn with_parameter_names<I: IntoIterator<Item = String>>(&mut self, names: I) -> &mut Self {
+        let names = names.into_iter().collect::<Vec<String>>();
+        assert!(names.len() == self.dimension);
+        self.parameter_names = Some(names);
+        self
+    }
+    /// Set the maximum number of steps to perform before failure (default: 4000).
+    pub fn with_max_steps(&mut self, max_steps: usize) -> &mut Self {
+        self.max_steps = max_steps;
+        self
+    }
     /// Set the [`AbortSignal`] of the [`Minimizer`].
     pub fn with_abort_signal(&mut self, abort_signal: Box<dyn AbortSignal>) -> &mut Self {
         self.abort_signal = abort_signal;
@@ -91,6 +124,22 @@ impl<S: Status, U: Default, E> Minimizer<S, U, E> {
         self.observers.push(observer);
         self
     }
+
+    /// Check parameters against the bounds
+    pub fn assert_parameters(&self, x: &[f64]) {
+        if let Some(bounds) = &self.bounds {
+            for (i, (x_i, bound_i)) in x.iter().zip(bounds).enumerate() {
+                assert!(
+                    bound_i.contains(*x_i),
+                    "Parameter #{} = {} is outside of the given bound: {}",
+                    i,
+                    x_i,
+                    bound_i
+                )
+            }
+        }
+    }
+
     /// Minimize the given [`CostFunction`] starting at the point `x0`.
     ///
     /// This method first runs [`Solver::initialize`], then runs [`Solver::step`] in a loop,
@@ -113,15 +162,19 @@ impl<S: Status, U: Default, E> Minimizer<S, U, E> {
     pub fn minimize(&mut self, func: &dyn CostFunction<U, E>) -> Result<(), E> {
         self.status.reset();
         self.abort_signal.reset();
-        self.solver
-            .initialize(func, &self.config, &mut self.status, &mut self.user_data)?;
+        self.solver.initialize(
+            func,
+            self.bounds.as_ref(),
+            &mut self.status,
+            &mut self.user_data,
+        )?;
         let mut current_step = 0;
         let mut observer_termination = false;
-        while current_step <= self.config.max_steps
+        while current_step <= self.max_steps
             && !observer_termination
             && !self.solver.check_for_termination(
                 func,
-                &self.config,
+                self.bounds.as_ref(),
                 &mut self.status,
                 &mut self.user_data,
             )?
@@ -130,7 +183,7 @@ impl<S: Status, U: Default, E> Minimizer<S, U, E> {
             self.solver.step(
                 current_step,
                 func,
-                &self.config,
+                self.bounds.as_ref(),
                 &mut self.status,
                 &mut self.user_data,
             )?;
@@ -139,15 +192,20 @@ impl<S: Status, U: Default, E> Minimizer<S, U, E> {
                 for observer in self.observers.iter_mut() {
                     observer_termination = observer.write().callback(
                         current_step,
+                        self.bounds.as_ref(),
                         &mut self.status,
                         &mut self.user_data,
                     ) || observer_termination;
                 }
             }
         }
-        self.solver
-            .postprocessing(func, &self.config, &mut self.status, &mut self.user_data)?;
-        if current_step > self.config.max_steps && !self.status.converged() {
+        self.solver.postprocessing(
+            func,
+            self.bounds.as_ref(),
+            &mut self.status,
+            &mut self.user_data,
+        )?;
+        if current_step > self.max_steps && !self.status.converged() {
             self.status.update_message("MAX EVALS");
         }
         if self.abort_signal.is_aborted() {
@@ -155,7 +213,13 @@ impl<S: Status, U: Default, E> Minimizer<S, U, E> {
         }
         self.result = self
             .solver
-            .summarize(func, &self.config, &self.status, &self.user_data)
+            .summarize(
+                func,
+                self.bounds.as_ref(),
+                self.parameter_names.as_ref(),
+                &self.status,
+                &self.user_data,
+            )
             .ok();
         Ok(())
     }
