+++ src/core/mod.rs
@@ -2,8 +2,6 @@
 pub mod abort_signal;
 /// [`Bound`] type for binding variables to a range.
 pub mod bound;
-/// [`Config`] type for configuring the minimizer.
-pub mod config;
 /// [`Minimizer`] type for the minimization process.
 pub mod minimizer;
 /// [`Point`] type for defining a point in the parameter space.
@@ -13,7 +11,6 @@ pub mod summary;
 
 pub use abort_signal::{AtomicAbortSignal, CtrlCAbortSignal, NopAbortSignal};
 pub use bound::Bound;
-pub use config::Config;
 pub use minimizer::Minimizer;
 pub use point::Point;
 pub use summary::Summary;
