+++ src/solvers/gradient_free/nelder_mead.rs
@@ -3,7 +3,7 @@ use std::fmt::Debug;
 use nalgebra::{DMatrix, DVector};
 
 use crate::{
-    core::{Bound, Config, Point, Summary},
+    core::{Bound, Point, Summary},
     traits::{CostFunction, Hessian, Solver},
     Float,
 };
@@ -586,17 +586,14 @@ impl<U, E> Solver<GradientFreeStatus, U, E> for NelderMead {
     fn initialize(
         &mut self,
         func: &dyn CostFunction<U, E>,
-        config: &Config,
+        bounds: Option<&Vec<Bound>>,
         status: &mut GradientFreeStatus,
         user_data: &mut U,
     ) -> Result<(), E> {
-        self.simplex = self.construction_method.generate(
-            func,
-            status.x0.as_slice(),
-            config.bounds.as_ref(),
-            user_data,
-        )?;
-        status.with_position(self.simplex.best_position(config.bounds.as_ref()));
+        self.simplex =
+            self.construction_method
+                .generate(func, status.x0.as_slice(), bounds, user_data)?;
+        status.with_position(self.simplex.best_position(bounds));
         Ok(())
     }
 
@@ -604,7 +601,7 @@ impl<U, E> Solver<GradientFreeStatus, U, E> for NelderMead {
         &mut self,
         _i_step: usize,
         func: &dyn CostFunction<U, E>,
-        config: &Config,
+        bounds: Option<&Vec<Bound>>,
         status: &mut GradientFreeStatus,
         user_data: &mut U,
     ) -> Result<(), E> {
@@ -613,7 +610,7 @@ impl<U, E> Solver<GradientFreeStatus, U, E> for NelderMead {
         let l = self.simplex.best();
         let c = &self.simplex.centroid;
         let mut xr = Point::from(c + (c - &h.x).scale(self.alpha));
-        xr.evaluate_bounded(func, config.bounds.as_ref(), user_data)?;
+        xr.evaluate_bounded(func, bounds, user_data)?;
         status.inc_n_f_evals();
         if l <= &xr && &xr < s {
             // Reflect if l <= x_r < s
@@ -621,7 +618,7 @@ impl<U, E> Solver<GradientFreeStatus, U, E> for NelderMead {
             // it should go. We have to do a sort, but it should be quick since most of the simplex
             // is already sorted.
             self.simplex.insert_and_sort(self.simplex.dimension - 2, xr);
-            status.with_position(self.simplex.best_position(config.bounds.as_ref()));
+            status.with_position(self.simplex.best_position(bounds));
             status.with_message("REFLECT");
             self.simplex.scale_volume(self.alpha);
             return Ok(());
@@ -631,7 +628,7 @@ impl<U, E> Solver<GradientFreeStatus, U, E> for NelderMead {
             // accept the expanded point x_e regardless (greedy expansion), or we should do one
             // final comparison between x_r and x_e and choose the smallest (greedy minimization).
             let mut xe = Point::from(c + (&xr.x - c).scale(self.beta));
-            xe.evaluate_bounded(func, config.bounds.as_ref(), user_data)?;
+            xe.evaluate_bounded(func, bounds, user_data)?;
             status.inc_n_f_evals();
             self.simplex.insert_sorted(
                 0,
@@ -646,7 +643,7 @@ impl<U, E> Solver<GradientFreeStatus, U, E> for NelderMead {
                     SimplexExpansionMethod::GreedyExpansion => xe,
                 },
             );
-            status.with_position(self.simplex.best_position(config.bounds.as_ref()));
+            status.with_position(self.simplex.best_position(bounds));
             status.with_message("EXPAND");
             self.simplex.scale_volume(self.alpha * self.beta);
             return Ok(());
@@ -668,14 +665,14 @@ impl<U, E> Solver<GradientFreeStatus, U, E> for NelderMead {
             if &xr < h {
                 // Try to contract outside if x_r < h
                 let mut xc = Point::from(c + (&xr.x - c).scale(self.gamma));
-                xc.evaluate_bounded(func, config.bounds.as_ref(), user_data)?;
+                xc.evaluate_bounded(func, bounds, user_data)?;
                 status.inc_n_f_evals();
                 if xc <= xr {
                     if &xc < s {
                         // If we are better than the second-worst, we need to sort everything, we
                         // could technically be anywhere, even in a new best.
                         self.simplex.insert_and_sort(self.simplex.dimension - 1, xc);
-                        status.with_position(self.simplex.best_position(config.bounds.as_ref()));
+                        status.with_position(self.simplex.best_position(bounds));
                     } else {
                         // Otherwise, we don't even need to update the best position, this was just
                         // a new worst or equal to second worst.
@@ -689,14 +686,14 @@ impl<U, E> Solver<GradientFreeStatus, U, E> for NelderMead {
             } else {
                 // Contract inside if h <= x_r
                 let mut xc = Point::from(c + (&h.x - c).scale(self.gamma));
-                xc.evaluate_bounded(func, config.bounds.as_ref(), user_data)?;
+                xc.evaluate_bounded(func, bounds, user_data)?;
                 status.inc_n_f_evals();
                 if &xc < h {
                     if &xc < s {
                         // If we are better than the second-worst, we need to sort everything, we
                         // could technically be anywhere, even in a new best.
                         self.simplex.insert_and_sort(self.simplex.dimension - 1, xc);
-                        status.with_position(self.simplex.best_position(config.bounds.as_ref()));
+                        status.with_position(self.simplex.best_position(bounds));
                     } else {
                         // Otherwise, we don't even need to update the best position, this was just
                         // a new worst or equal to second worst.
@@ -712,7 +709,7 @@ impl<U, E> Solver<GradientFreeStatus, U, E> for NelderMead {
         let l_clone = l.clone();
         for p in self.simplex.points.iter_mut().skip(1) {
             *p = Point::from(&l_clone.x + (&p.x - &l_clone.x).scale(self.delta));
-            p.evaluate_bounded(func, config.bounds.as_ref(), user_data)?;
+            p.evaluate_bounded(func, bounds, user_data)?;
             status.inc_n_f_evals();
         }
         // We must do a fresh sort here, since we don't know the ordering of the shrunken simplex,
@@ -721,7 +718,7 @@ impl<U, E> Solver<GradientFreeStatus, U, E> for NelderMead {
         self.simplex.sort();
         // We also need to recalculate the centroid and figure out if there's a new best position:
         self.simplex.compute_centroid();
-        status.with_position(self.simplex.best_position(config.bounds.as_ref()));
+        status.with_position(self.simplex.best_position(bounds));
         status.with_message("SHRINK");
         self.simplex
             .scale_volume(Float::powi(self.delta, self.simplex.dimension as i32));
@@ -731,7 +728,7 @@ impl<U, E> Solver<GradientFreeStatus, U, E> for NelderMead {
     fn check_for_termination(
         &mut self,
         _func: &dyn CostFunction<U, E>,
-        _config: &Config,
+        _bounds: Option<&Vec<Bound>>,
         status: &mut GradientFreeStatus,
         _user_data: &mut U,
     ) -> Result<bool, E> {
@@ -751,7 +748,7 @@ impl<U, E> Solver<GradientFreeStatus, U, E> for NelderMead {
     fn postprocessing(
         &mut self,
         func: &dyn CostFunction<U, E>,
-        _config: &Config,
+        _bounds: Option<&Vec<Bound>>,
         status: &mut GradientFreeStatus,
         user_data: &mut U,
     ) -> Result<(), E> {
@@ -765,7 +762,8 @@ impl<U, E> Solver<GradientFreeStatus, U, E> for NelderMead {
     fn summarize(
         &self,
         _func: &dyn CostFunction<U, E>,
-        config: &Config,
+        bounds: Option<&Vec<Bound>>,
+        parameter_names: Option<&Vec<String>>,
         status: &GradientFreeStatus,
         _user_data: &U,
     ) -> Result<Summary, E> {
@@ -773,15 +771,12 @@ impl<U, E> Solver<GradientFreeStatus, U, E> for NelderMead {
             x0: status.x0.iter().cloned().collect(),
             x: status.x.iter().cloned().collect(),
             fx: status.fx,
-            bounds: config.bounds.clone(),
+            bounds: bounds.cloned(),
             converged: status.converged,
             cost_evals: status.n_f_evals,
             gradient_evals: 0,
             message: status.message.clone(),
-            parameter_names: config
-                .parameter_names
-                .as_ref()
-                .map(|names| names.iter().cloned().collect()),
+            parameter_names: parameter_names.map(|names| names.iter().cloned().collect()),
             std: status
                 .err
                 .as_ref()
@@ -840,7 +835,7 @@ mod tests {
     fn test_bounded_nelder_mead() -> Result<(), Infallible> {
         let algo = NelderMead::default();
         let mut m = Minimizer::new(Box::new(algo), 2).setup(|m| {
-            m.on_config(|c| c.with_bounds(vec![(-4.0, 4.0), (-4.0, 4.0)]))
+            m.with_bounds(vec![(-4.0, 4.0), (-4.0, 4.0)])
                 .with_abort_signal(CtrlCAbortSignal::new().boxed())
         });
         let problem = Rosenbrock { n: 2 };
